Current version of Psi+ is 1.5.1546
It is based on:
* psi: c0dc6cf8
* plugins: d6f8fc4
* psimedia: fda9f7c
* resources: 247f5dd

+++ iris/src/irisnet/noncore/stunallocate.cpp
@@ -427,7 +427,7 @@ public:
     DontFragmentState               dfState;
     QString                         clientSoftware, serverSoftware;
     QHostAddress                    reflexiveAddress, relayedAddress;
-    int                             reflexivePort, relayedPort;
+    quint16                         reflexivePort, relayedPort;
     StunMessage                     msg;
     int                             allocateLifetime;
     QTimer *                        allocateRefreshTimer;
@@ -1119,7 +1119,7 @@ int StunAllocate::reflexivePort() const { return d->reflexivePort; }
 
 QHostAddress StunAllocate::relayedAddress() const { return d->relayedAddress; }
 
-int StunAllocate::relayedPort() const { return d->relayedPort; }
+quint16 StunAllocate::relayedPort() const { return d->relayedPort; }
 
 QList<QHostAddress> StunAllocate::permissions() const { return d->permsOut; }
